feat: 注册时可设置昵称，新增个人信息编辑功能
- LoginFormDTO 添加 nickName 字段，注册时支持自定义昵称
- 新增 ProfileUpdateDTO 和 PUT /user/profile 接口保存个人信息
- 新增 ProfileEdit.vue 个人信息编辑页（昵称/城市/性别/生日/简介）
- HeaderBar 下拉菜单添加"个人信息"入口
- stores/user.js 初始化时自动拉取用户信息，修复刷新后用户状态丢失
- 升级 MyBatis-Plus 3.4.3 → 3.5.3.1 修复 Java 17+ 兼容性
- 添加 Jackson 日期序列化配置，LocalDate 输出为字符串而非数组
- 修复 count() 返回类型 Long 与 int 不匹配的编译错误

+++ backend/src/main/java/com/medqueue/controller/FollowController.java
@@ -51,7 +51,7 @@ public class FollowController {
     @GetMapping("/or/{id}/{type}")
     public Result isFollow(@PathVariable("id") Long followId, @PathVariable("type") Integer followType) {
         UserDTO user = UserHolder.getUser();
-        int count = followService.query()
+        long count = followService.query()
                 .eq("user_id", user.getId())
                 .eq("follow_id", followId)
                 .eq("follow_type", followType)
